added editing answers

+++ dev/admin/accueil.php
@@ -2,35 +2,39 @@
 
 if(isset($_SESSION['admin'])):
 
-
-		$sql = "SELECT idquestion, question, date, online, id_auteur, objet
-						FROM question
+	$sql = "SELECT idquestion, question, date, question.online, id_auteur, objet, idauteurs, nom, prenom
+						FROM question join auteurs
+						ON question.id_auteur = auteurs.idauteurs
 						ORDER BY date DESC";
 
 		$result = $connect->query($sql);
 		echo $connect->error;
-
-echo '<div class="login-form">';
-
-	echo '<a href="log.php?logout=yes">se déconnecter</a>
-				</div>
-				</header>
-				<main>
-				<p>votre niveau de droits: '.$_SESSION['admin']->niveau.'</p>';
-
-				if($result->num_rows > 0):
-					echo'<table class="admin-tableau">';
-
-					while($row = $result->fetch_object()):
-						echo '<tr>
-									<td>'.$row->objet.'</td>
-									<td><a href="question.php?edit='.$row->idquestion.'">Éditer</a></td>
-									<td><a href="question.php?delete='.$row->idquestion.'">Supprimer</a></td>
-									</tr>';
-
-					endwhile;	
-
-					echo '</table>';
-				endif;
-endif;
 ?>
+
+<p class="rights-level">votre niveau de droits: <?php echo $_SESSION['admin']->niveau; ?></p>
+
+				<?php if($result->num_rows > 0): ?>
+					<table class="admin-tableau">
+						<thead>
+							<th>objet</th>
+							<th>auteur</th>
+							<th>date</th>
+							<th>Éditer</th>
+							<th>Réponses</th>
+							<th>Effacer</th>
+						</thead>
+					<?php while($row = $result->fetch_object()): ?>
+
+						<tr>
+							<td><?php echo $row->objet ?></td>
+							<td><?php echo $row->prenom." ".$row->nom ?></td>
+							<td><?php echo $row->objet ?></td>
+							<td><a href="index.php?editQ=<?php echo $row->idquestion ?>">Éditer</a></td>
+							<td><a href="index.php?reponses=<?php echo $row->idquestion?>">Réponses</a></td>
+							<td><a href="question.php?delete=<?php echo $row->idquestion ?>" class="btn--delete" onclick="if(!confirm(\'Sûr?\')) return false;">Supprimer</a></td>
+						</tr>
+
+		<?php endwhile;	?>
+
+					</table>
+<?php endif; endif; ?>
